Pickup buildable
\/ this doesnt matter unless ur dan \/

I'll do work outside of class to make up for all the time wasted trying to update zombie's attack to be better (I did it, but it was fundamentally different from before and in a way it was worse than the original version). I was trying to make it harder to dodge the zombie but it got too complex, we should just focus on radio defence and not put too much into player survival. RIP 4 hours of work.